refactor: add new test suite, move test to different test suites, rename test
