+++ imagetest/test_suites/license/package_license_test.go
@@ -1,6 +1,6 @@
 // +build cit
 
-package imagevalidation
+package license
 
 import (
 	"fmt"
