+++ imagetest/test_suites/service/clock_test.go
@@ -1,6 +1,6 @@
 // +build cit
 
-package imagevalidation
+package service
 
 import (
 	"io/ioutil"
