+++ imagetest/test_suites/service/setup.go
@@ -1,0 +1,16 @@
+package service
+
+import "github.com/GoogleCloudPlatform/guest-test-infra/imagetest"
+
+// Name is the name of the test package. It must match the directory name.
+var Name = "image_validation"
+
+// TestSetup sets up the test workflow.
+func TestSetup(t *imagetest.TestWorkflow) error {
+	_, err := t.CreateTestVM("vm1")
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
